+++ include/relu_layer.h
@@ -7,5 +7,5 @@ class relu_layer : public fc_layer
 	void compute_derivatives(const fc_layer& next) override;
 
 public:
-	explicit relu_layer(size_t prev_size, size_t size);
+	explicit relu_layer(int prev_size, int size);
 };
